# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "User API documentation",
        "contact": {}
    },
    "basePath": "/",
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Signup",
                        "name": "loginRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.BaseRes"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/signup": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Register an account",
                "parameters": [
                    {
                        "description": "Signup",
                        "name": "signupRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.SignupReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.BaseRes"
                        }
                    }
                }
            }
        },
        "/api/v1/words/list": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Core"
                ],
                "summary": "List words by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ids seperated  by commas, eg 1,2,3,4",
                        "name": "ids",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core.SearchRes"
                        }
                    }
                }
            }
        },
        "/api/v1/words/pronounce/{id}": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Core"
                ],
                "summary": "Add sound of pronunciation",
                "parameters": [
                    {
                        "type": "file",
                        "description": "this is a test file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id of word to upload pronounce",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core.BaseRes"
                        }
                    }
                }
            }
        },
        "/api/v1/words/search": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Core"
                ],
                "summary": "Search word in language",
                "parameters": [
                    {
                        "type": "string",
                        "description": "prefix of words",
                        "name": "prefix",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "keyword to search",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search word in language",
                        "name": "language",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "sort by alphabet asc/ desc, default asc",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core.SearchRes"
                        }
                    }
                }
            }
        },
        "/api/v1/words/translate": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Core"
                ],
                "summary": "Translate from Kinh to another language",
                "parameters": [
                    {
                        "type": "string",
                        "description": "keyword in Kinh",
                        "name": "keyword",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "language want to translate into",
                        "name": "language",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core.BaseRes"
                        }
                    }
                }
            }
        },
        "/api/v1/words/upsert": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Core"
                ],
                "summary": "Add or update word",
                "parameters": [
                    {
                        "description": "if id = 0 means create new word, else update word",
                        "name": "upsertWordReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/core.UpsertWord"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core.BaseRes"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.BaseRes": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "auth.LoginReq": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "string": {
                    "type": "string"
                }
            }
        },
        "auth.SignupReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "core.BaseRes": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "core.Definition": {
            "type": "object",
            "properties": {
                "examples": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "meaning": {
                    "type": "string"
                }
            }
        },
        "core.SearchRes": {
            "type": "object",
            "properties": {
                "words": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.WordRes"
                    }
                }
            }
        },
        "core.UpsertWord": {
            "type": "object",
            "properties": {
                "definitions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.Definition"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "language": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "word": {
                    "type": "string"
                }
            }
        },
        "core.WordRes": {
            "type": "object",
            "properties": {
                "definitions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.Definition"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "language": {
                    "type": "string"
                },
                "pronunciation_sound": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "word": {
                    "type": "string"
                }
            }
        }
    }
}